Enabled type coercion for all simple type properties

+++ tests/Nemo.Test/Program.cs
@@ -149,7 +149,7 @@ namespace NemoTest
 
             var retrieve_many_customer_sql = ObjectFactory.Retrieve<ICustomer>(sql: "select * from Customers where CustomerID in (@CustomerIDs)", parameters: new { CustomerIDs = new[] { "ALFKI", "ANTON" } });
 
-            var retrieve_customer_orders_ids_sql = ObjectFactory.Retrieve<int>(sql: "select OrderID from Orders where CustomerID = @CustomerID", parameters: new ParamList { CustomerID => "ALFKI" });
+            var retrieve_customer_orders_ids_sql = ObjectFactory.Retrieve<long>(sql: "select OrderID from Orders where CustomerID = @CustomerID", parameters: new ParamList { CustomerID => "ALFKI" });
 
             // Advanced!
             // Retrieve customers with orders as object graph
